Update dependencies to newer versions
In this commit, the dependencies "go-cursor", "go-pretty", and "go/sys" of the "go-input" module were updated to their latest versions. This was done to incorporate the latest features and bug fixes that these dependencies provide. Also, the Go version was incremented from 1.21.0 to 1.21.3, and the toolchain was set to go1.21.4.