Add password input functionality
Added a function to the input.go file that allows for password input functionality. This function, 'GetPassword' prompts for user password without displaying it in plaintext for security measures. For this update, the Go libraries 'crypto' and 'terminal' were added to the go.mod and go.sum files to assist with secure user password entry.